requirements.txt
-i https://pypi.org/simple
django==2.2
djangorestframework==3.11.1
easyprocess==0.3
numpy==1.19.1; python_version >= '3.6'
pandas==1.1.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
pyvirtualdisplay==1.3.2
selenium==3.141.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
